+++ back-end/api/app/Models/Appliance.php
@@ -9,8 +9,11 @@ class Appliance extends Model
 {
     use HasFactory;
 
+    protected $guarded = [
+        'id'
+    ];
+
     protected $fillable = [
-        'id',
         'name',
         'description',
         'voltage',
@@ -23,4 +26,6 @@ class Appliance extends Model
         'voltage' => 'string',
         'brand' => 'string'
     ];
+
+    public $timestamps = false;
 }