+++ back-end/api/app/UseCases/IService.php
@@ -1,9 +1,0 @@
-<?php
-namespace App\Implementations;
-
-interface IService{
-
-    function __construct($repository);
-    public function execute($data);
-
-}