+++ back-end/api/routes/Appliance/ApplianceCreateRoute.php
@@ -1,8 +1,0 @@
-<?php
-
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Route;
-
-Route::get("/appliance",function (Request $request){
-    return response()->json("Hello World!");
-});