+++ CHANGELOG.md
@@ -15,9 +15,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add version 1.0.0 of diagram [TTFS-4](https://guilhermetelesdev.atlassian.net/browse/TTFS-4)
 
+- Add CRUD back end [TTFS-8](https://guilhermetelesdev.atlassian.net/browse/TTFS-8)
+
+- Add basic interface form front end [TTFS-9](https://guilhermetelesdev.atlassian.net/browse/TTFS-9)
+
+- Documentation done [TTFS-7](https://guilhermetelesdev.atlassian.net/browse/TTFS-7)
+
 ### Fixed
 
-- 
+- change structure of CRUD back end [TTFS-8](https://guilhermetelesdev.atlassian.net/browse/TTFS-8)
 
 ### Removed